src/aggregator-lambda/requirements.txt
boto3
azure-storage-blob
